+++ src/components/ndn-lib/tests/local_signal_ndn_mgr_obj.rs
@@ -21,15 +21,6 @@ fn generate_random_bytes(size: u64) -> Vec<u8> {
     buffer
 }
 
-fn generate_random_chunk(size: u64) -> (ChunkId, Vec<u8>) {
-    let chunk_data = generate_random_bytes(size);
-    let mut hasher = ChunkHasher::new(None).expect("hash failed.");
-    let hash = hasher.calc_from_bytes(&chunk_data);
-    let chunk_id = ChunkId::from_sha256_result(&hash);
-    info!("chunk_id: {}", chunk_id.to_string());
-    (chunk_id, chunk_data)
-}
-
 fn generate_random_obj() -> (ObjId, serde_json::Value) {
     let int_value = rand::random::<u32>();
     let str_value: String = generate_random_bytes(7).encode_hex();
@@ -46,35 +37,6 @@ fn generate_random_obj() -> (ObjId, serde_json::Value) {
     });
     let (obj_id, _obj_str) = build_named_object_by_json("non-test-obj", &test_obj);
     (obj_id, test_obj)
-}
-
-async fn write_chunk(ndn_mgr_id: &str, chunk_id: &ChunkId, chunk_data: &[u8]) {
-    let (mut chunk_writer, _progress_info) =
-        NamedDataMgr::open_chunk_writer(Some(ndn_mgr_id), chunk_id, chunk_data.len() as u64, 0)
-            .await
-            .expect("open chunk writer failed");
-    chunk_writer
-        .write_all(chunk_data)
-        .await
-        .expect("write chunk to ndn-mgr failed");
-    NamedDataMgr::complete_chunk_writer(Some(ndn_mgr_id), chunk_id)
-        .await
-        .expect("wait chunk writer complete failed.");
-}
-
-async fn read_chunk(ndn_mgr_id: &str, chunk_id: &ChunkId) -> Vec<u8> {
-    let (mut chunk_reader, len) =
-        NamedDataMgr::open_chunk_reader(Some(ndn_mgr_id), chunk_id, SeekFrom::Start(0), false)
-            .await
-            .expect("open reader from ndn-mgr failed.");
-
-    let mut buffer = vec![0u8; len as usize];
-    chunk_reader
-        .read_exact(&mut buffer)
-        .await
-        .expect("read chunk from ndn-mgr failed");
-
-    buffer
 }
 
 async fn check_obj_inner_path(
@@ -245,79 +207,6 @@ async fn init_ndn_server(ndn_mgr_id: &str) -> (NdnClient, NdnServerHost) {
     );
 
     (client, host)
-}
-
-#[tokio::test]
-async fn ndn_local_chunk_ok() {
-    init_logging("ndn_local_chunk_ok", false);
-
-    let ndn_mgr_id: String = generate_random_bytes(16).encode_hex();
-    let _ndn_client = init_ndn_server(ndn_mgr_id.as_str()).await;
-
-    let chunk_size: u64 = 1024 * 1024 + 515;
-    let (chunk_id, chunk_data) = generate_random_chunk(chunk_size);
-
-    write_chunk(ndn_mgr_id.as_str(), &chunk_id, chunk_data.as_slice()).await;
-
-    let buffer = read_chunk(ndn_mgr_id.as_str(), &chunk_id).await;
-
-    assert_eq!(buffer, chunk_data, "chunk-content check failed");
-}
-
-#[tokio::test]
-async fn ndn_local_chunk_not_found() {
-    init_logging("ndn_local_chunk_not_found", false);
-
-    let ndn_mgr_id: String = generate_random_bytes(16).encode_hex();
-    let _ndn_client = init_ndn_server(ndn_mgr_id.as_str()).await;
-
-    let chunk_size: u64 = 1024 + 154;
-    let (chunk_id, chunk_data) = generate_random_chunk(chunk_size);
-
-    // Pull the chunk using the NdnClient
-    let ret = NamedDataMgr::open_chunk_reader(
-        Some(ndn_mgr_id.as_str()),
-        &chunk_id,
-        SeekFrom::Start(0),
-        false,
-    )
-    .await;
-
-    match ret {
-        Ok(_) => assert!(false, "should no chunk found"),
-        Err(err) => match err {
-            NdnError::NotFound(_) => {
-                info!("Chunk not found as expected");
-            }
-            _ => {
-                assert!(false, "Unexpected error type");
-            }
-        },
-    }
-}
-
-#[tokio::test]
-async fn ndn_local_chunk_verify_failed() {
-    init_logging("ndn_local_chunk_verify_failed", false);
-
-    let ndn_mgr_id: String = generate_random_bytes(16).encode_hex();
-    let _ndn_client = init_ndn_server(ndn_mgr_id.as_str()).await;
-
-    let chunk_size: u64 = 1024 * 1024 + 567;
-    let (chunk_id, chunk_data) = generate_random_chunk(chunk_size);
-    let mut fake_chunk_data = chunk_data.clone();
-    fake_chunk_data.splice(0..10, 0..10);
-
-    write_chunk(ndn_mgr_id.as_str(), &chunk_id, &fake_chunk_data).await;
-
-    let buffer = read_chunk(ndn_mgr_id.as_str(), &chunk_id).await;
-
-    assert_eq!(buffer, fake_chunk_data, "chunk-content check failed");
-
-    let mut hasher = ChunkHasher::new(None).expect("hash failed.");
-    let hash = hasher.calc_from_bytes(&buffer);
-    let fake_chunk_id = ChunkId::from_sha256_result(&hash);
-    assert_ne!(fake_chunk_id, chunk_id, "chunk-id should mismatch");
 }
 
 #[tokio::test]
